requirements.txt
fastapi==0.79.0
mysql_connector_repackaged==0.3.1
pydantic==1.9.1
PyJWT==2.4.0
cryptography==36.0.1
python-dotenv==0.20.0
uvicorn==0.18.2
mysql-connector-python
aiofiles==0.6.0
python-multipart==0.0.5
jinja2==2.11.2
gunicorn
uvloop
httptools